task 2

+++ pagination/2-hypermedia_pagination.py
@@ -66,7 +66,7 @@ class Server:
             hyper["data"] = self.get_page(page, page_size)
             if not hyper["data"]:
                 hyper["page_size"] = 0
-            hyper["next_page"] = None if page > (len(f_read) / page_size) else page + 1
+            hyper["next_page"] = None if page > int(len(f_read) / page_size) else page + 1
             hyper["prev_page"] = None if page < 2 else page - 1
             hyper["total_pages"] = int(len(f_read) / page_size)
